Codewars: Fixup YouAreASquare exercise

+++ codewars/src/main/scala/YouAreASquare.scala
@@ -1,11 +1,6 @@
 object YouAreASquare {
   def isSquare(potentialSquare: Int): Boolean = {
-    val potentialRoot = math.sqrt(potentialSquare)
+    val potentialRoot = math.sqrt(potentialSquare).toInt
     potentialSquare.toDouble == math.pow(potentialRoot, 2)
   }
 }
-
-//isSquare(154) should return false
-//88 118
-//Test Failed
-//  true was not false
